convert entity manager factory to ts
