+++ GreyTide/GreyTide.csproj
@@ -247,7 +247,6 @@
     <Content Include="app\tide\item.html" />
     <Content Include="app\tide\items.html" />
     <Content Include="packages.config" />
-    <Content Include="app\greyTideContext.js" />
     <Content Include="scripts\angular-animate.js" />
     <Content Include="scripts\angular-animate.min.js" />
     <Content Include="scripts\angular-mocks.js" />
@@ -363,8 +362,9 @@
     <TypeScriptCompile Include="app\directives\cc-widget-minimize.ts" />
     <TypeScriptCompile Include="app\layout\shellCtrl.ts" />
     <TypeScriptCompile Include="app\layout\sidebarCtrl.ts" />
-    <TypeScriptCompile Include="app\services\datacontext.ts" />
     <TypeScriptCompile Include="app\services\entityManagerFactory.ts" />
+    <TypeScriptCompile Include="app\services\guid.ts" />
+    <TypeScriptCompile Include="app\services\datacontext.ts" />
     <TypeScriptCompile Include="app\state\state.ts" />
     <TypeScriptCompile Include="app\tide\tide.ts" />
     <TypeScriptCompile Include="scripts\typings\angularjs\angular-animate.d.ts" />
@@ -382,6 +382,7 @@
     <TypeScriptCompile Include="scripts\typings\linq\linq.d.ts" />
     <TypeScriptCompile Include="scripts\typings\linq\linq.jquery.d.ts" />
     <TypeScriptCompile Include="scripts\typings\q\Q.d.ts" />
+    <TypeScriptCompile Include="scripts\typings\statemachine\statemachine.d.ts" />
     <TypeScriptCompile Include="scripts\typings\toastr\toastr.d.ts" />
   </ItemGroup>
   <ItemGroup>
