+++ GreyTide/app/services/guid.ts
@@ -1,0 +1,15 @@
+﻿/// <reference path="../../scripts/typings/angularjs/angular.d.ts" />
+/// <reference path="../../scripts/typings/breeze/breeze.d.ts" />
+
+module App.Services.GuidGenerator {
+    export function newGuid() {
+        function s4() {
+            return Math.floor((1 + Math.random()) * 0x10000)
+                .toString(16)
+                .substring(1);
+        }
+        return s4() + s4() + '-' + s4() + '-' + s4() + '-' +
+            s4() + '-' + s4() + s4() + s4();
+    }
+
+}
