+++ GreyTide/index.html
@@ -112,6 +112,7 @@
     <script src="scripts/multi-select/angular-multi-select.js"></script>
     
     <!-- app Services -->
+    <script src="app/services/guid.js"></script>
     <script src="app/services/datacontext.js"></script>
     <script src="app/services/entityManagerFactory.js"></script>
 </body>
