+++ GreyTide/scripts/typings/statemachine/statemachine.d.ts
@@ -1,0 +1,5 @@
+﻿declare var StateMachine: any;
+
+declare module "state-machine" {
+    export = StateMachine;
+}
